- Add cancelOrVoid method in OrderService - Create OrderValidationService - Create validateCancelable method in OrderValidationService - Create StockMutationService - Create buildReturnStockMutations method in StockMutationService - Create applyReturnedStockMutation method in StockMutationService - Create cancelOrVoid method in OrderController - Create findByReferenceId repo method in StockMutationRepository - Change stock and quantity properties in Product and OrderItems entities

+++ src/main/java/com/gonku/pos_be/dto/product/ProductRequestDto.java
@@ -30,7 +30,7 @@ public class ProductRequestDto {
     private BigDecimal salePrice;
 
     @Min(value = 0, message = "Stock must be equal or greater than 0")
-    private Integer stock;
+    private BigDecimal stock;
 
     private String picturePath;
